PythonScraper/requirements.txt
# BetSnipe.ai v2.0 Requirements
# Install with: pip install -r requirements.txt

# ==========================================
# Core Dependencies
# ==========================================

# HTTP clients
aiohttp>=3.9.0
requests>=2.31.0
cloudscraper>=1.2.71

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Fuzzy matching
rapidfuzz>=3.0.0

# Telegram bot
python-telegram-bot>=20.0

# Environment variables
python-dotenv>=1.0.0

# HTML parsing (used by some scrapers)
beautifulsoup4>=4.12.0

# ==========================================
# NEW v2.0 Dependencies
# ==========================================

# Async PostgreSQL database
asyncpg>=0.29.0

# FastAPI web framework
fastapi>=0.109.0

# ASGI server
uvicorn[standard]>=0.27.0

# WebSocket support
websockets>=12.0

# Configuration management
pydantic>=2.0.0
pydantic-settings>=2.1.0

# ==========================================
# Optional Dependencies
# ==========================================

# SQL Server support (legacy)
# pyodbc>=5.0.0

# Supabase client (alternative to direct asyncpg)
# supabase>=2.0.0

# Testing
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
# httpx>=0.25.0  # For testing FastAPI

# Development
# black>=23.0.0
# isort>=5.12.0
# mypy>=1.0.0
